Eliminação de "tokens" desnecessários e correção no tratamento do prompt
