+++ src/chat/chatTools.ts
@@ -1,29 +1,18 @@
-import path from "path";
 import * as vscode from "vscode";
-import * as fs from "fs";
-import { sendPatchApplyRequest, sendPatchValidateRequest, ValidResponse } from "../protocolMessages";
-import { ServersConfig } from "../utils";
 import {
 	APPLY_PATCH_COMMAND, COMPILER_COMMAND, COMPILER_PARTICIPANT_ID, COMPILER_TOOL_NAME,
-	ChatCompilerToolInput, DIAGNOSTIC_IDLE_WINDOW_MS, DIAGNOSTIC_WAIT_TIMEOUT_MS,
-	DIAGNOSTIC_WAIT_TIMEOUT_MS_FOLDER, DiagnosticEntry, DiagnosticsWaitResult,
-	FlagOptions, HELP_COMMAND, MAX_DIAGNOSTICS_TO_SHOW, REBUILD_FILE_COMMAND,
-	REBUILD_OPEN_EDITORS_COMMAND, REBUILD_WORKSPACE_COMMAND, SYNTAX_ONLY_COMMAND,
-	SYNTAX_ONLY_FILE_COMMAND, VALIDATE_PATCH_COMMAND, applyDiagnosticFilterOptions,
-	collectDiagnostics, formatDiagnosticsSummary, isCurrentEditorTargetAlias,
-	isDirectoryUri, isOpenEditorsTargetAlias, isWorkspaceTargetAlias, normalizeCommand,
-	normalizeTargetKeyword, resolvePatchFilePath, resolveTarget, resolveTargetUri,
+	ChatCompilerToolInput, DEFAULT_COMPILER_TOOL_INPUT, DEFAULT_FLAG_OPTIONS, DIAGNOSTIC_IDLE_WINDOW_MS, DIAGNOSTIC_WAIT_TIMEOUT_MS,
+	DiagnosticEntry, DiagnosticsWaitResult,
+	FlagOptions, HELP_COMMAND,
+	SYNTAX_ONLY_COMMAND,
+	VALIDATE_PATCH_COMMAND,
+	isCurrentEditorTargetAlias,
+	isOpenEditorsTargetAlias, isWorkspaceTargetAlias, normalizeCommand,
+	normalizeTargetKeyword, resolveTarget,
 	stripQuotes
 } from "./chatUtils";
-
-const DEFAULT_FLAG_OPTIONS: FlagOptions = {
-	only: "all",
-	sort: "none",
-	format: "markdown",
-	syntaxOnly: false,
-	applyOld: false,
-	applied: []
-};
+import { handleApplyPatchCommand, handleValidPatchCommand } from "./handlePatchCommands";
+import { handleCompilerCommand } from "./handleCompileCommands";
 
 function isValidCommand(command: string | undefined, prompt: string | undefined): string {
 	let result: string = "";
@@ -91,7 +80,7 @@ function extractTargetFromPrompt(prompt: string): string | undefined {
 		return stripQuotes(directiveMatch[1]);
 	}
 
-	const normalizedValue: string = normalizeTargetKeyword(stripQuotes(prompt));
+	const normalizedValue: string = normalizeTargetKeyword(prompt);
 	if (
 		isWorkspaceTargetAlias(normalizedValue) ||
 		isCurrentEditorTargetAlias(normalizedValue) ||
@@ -115,245 +104,47 @@ function extractTargetFromPrompt(prompt: string): string | undefined {
 }
 
 /**
- * Extracts patch file path from a standalone prompt value.
- * Supported input examples: relative path, absolute path, or quoted path.
- * @param prompt Full prompt sent in the chat.
- * @returns Extracted patch file path when found.
- */
-function extractPatchFileFromPrompt(prompt: string): string | undefined {
-	const value: string = stripQuotes(prompt);
-	if (!value) {
-		return undefined;
-	}
-
-	const quotedPrompt: string = prompt.trim();
-	if (
-		(quotedPrompt.startsWith("\"") && quotedPrompt.endsWith("\"")) ||
-		(quotedPrompt.startsWith("'") && quotedPrompt.endsWith("'"))
-	) {
-		return value;
-	}
-
-	if (!value.includes(" ")) {
-		return value;
-	}
-
-	return undefined;
-}
-
-/**
  * Parses diagnostic filter/sort flags from the prompt.
  * @param flags Raw list of flags received by the tool.
  * @returns Normalized filter/sort options and the list of applied flags.
  */
-function parseFlagOptions(flags: string[] | undefined): FlagOptions {
-	if (!flags || flags.length === 0) {
-		return { ...DEFAULT_FLAG_OPTIONS, applied: [] };
-	}
+function parseFlagOptions(options: FlagOptions, key: string, value: string): void {
+	key = key.toLowerCase();
+	value = stripQuotes(value).toLowerCase();
 
-	const joined: string = flags.join(" ");
-	const options: FlagOptions = {
-		...DEFAULT_FLAG_OPTIONS,
-		applied: []
-	};
-	const regex: RegExp = /(syntaxOnly|only|max|sort|format|applyOld)\s*[:=]\s*("[^"]+"|"[^"]+"|\S+)/gi;
-	let match: RegExpExecArray | null;
-
-	while ((match = regex.exec(joined)) !== null) {
-		const key: string = match[1].toLowerCase();
-		const rawValue: string = stripQuotes(match[2]);
-		const value: string = rawValue.toLowerCase();
-
-		if (key === "syntaxonly") {
-			options.syntaxOnly = value === "true";
-			options.applied.push("syntaxOnly=true");
+	if (key === "syntaxonly") {
+		options.syntaxOnly = value === "true";
+		options.applied.push("syntaxOnly=true");
+	} else if (key === "applyold") {
+		options.applyOld = value === "true";
+		options.applied.push("applyOld=true");
+	} else if (key === "only") {
+		if (value === "error" || value === "errors") {
+			options.only = "error";
+			options.applied.push("only=error");
+		} else if (value === "warning" || value === "warnings" || value === "warn") {
+			options.only = "warning";
+			options.applied.push("only=warning");
+		} else if (value === "all") {
+			options.only = "all";
+			options.applied.push("only=all");
 		}
-
-		if (key === "applyold") {
-			options.applyOld = value === "true";
-			options.applied.push("applyOld=true");
+	} else if (key === "max") {
+		const max: number = Number.parseInt(value, 10);
+		if (!Number.isNaN(max) && max > 0) {
+			options.max = max;
+			options.applied.push(`max=${max}`);
 		}
-
-		if (key === "only") {
-			if (value === "error" || value === "errors") {
-				options.only = "error";
-				options.applied.push("only=error");
-			} else if (value === "warning" || value === "warnings" || value === "warn") {
-				options.only = "warning";
-				options.applied.push("only=warning");
-			} else if (value === "all") {
-				options.only = "all";
-				options.applied.push("only=all");
-			}
+	} else if (key === "sort") {
+		if (value === "file" || value === "severity" || value === "none") {
+			options.sort = value;
+			options.applied.push(`sort=${value}`);
 		}
-
-		if (key === "max") {
-			const max: number = Number.parseInt(value, 10);
-			if (!Number.isNaN(max) && max > 0) {
-				options.max = max;
-				options.applied.push(`max=${max}`);
-			}
+	} else if (key === "format") {
+		if (value === "markdown" || value === "json") {
+			options.format = value;
+			options.applied.push(`format=${value}`);
 		}
-
-		if (key === "sort") {
-			if (value === "file" || value === "severity" || value === "none") {
-				options.sort = value;
-				options.applied.push(`sort=${value}`);
-			}
-		}
-
-		if (key === "format") {
-			if (value === "markdown" || value === "json") {
-				options.format = value;
-				options.applied.push(`format=${value}`);
-			}
-		}
-	}
-
-	return options;
-}
-
-/**
- * Validates if a file extension is supported by patch validation flow.
- * @param filePath Candidate file path.
- * @returns True when extension is accepted for patch validation.
- */
-function isSupportedPatchFile(filePath: string): boolean {
-	const extension: string = path.extname(filePath).toLowerCase();
-	return extension === ".ptm" || extension === ".zip" || extension === ".upd";
-}
-
-/**
- * Runs patch validation command from chat.
- * Validates prompt/path/extension and then requests server-side validation.
- * @param prompt Raw prompt text.
- * @param token Request cancellation token.
- * @param isError Mutable wrapper used as out parameter.
- * The function updates `isError.value` to indicate whether the validation flow ended with an error.
- * @returns Status text for chat output.
- */
-async function handleValidPatchCommand(
-	prompt: string | undefined,
-	token: vscode.CancellationToken,
-	isError: { value: boolean }
-): Promise<string> {
-	isError.value = true;
-
-	if (token.isCancellationRequested) {
-		return "Validate patch command canceled by user.";
-	}
-
-	const patchFromPrompt: string | undefined = prompt?.trim() ? extractPatchFileFromPrompt(prompt.trim()) : undefined;
-	if (!patchFromPrompt) {
-		return "Patch file is required. Usage: /validate-patch patches/fix.ptm or /validate-patch c:/patches/fix.zip";
-	}
-
-	const patchFilePath: string | undefined = resolvePatchFilePath(patchFromPrompt);
-	if (!patchFilePath) {
-		return `Unable to resolve patch file path: ${patchFromPrompt}`;
-	}
-
-	if (!fs.existsSync(patchFilePath)) {
-		return `Patch file not found: ${patchFilePath}`;
-	}
-
-	let fileStats: fs.Stats;
-	try {
-		fileStats = fs.statSync(patchFilePath);
-	} catch (error) {
-		return `Unable to access patch file: ${error instanceof Error ? error.message : String(error)}`;
-	}
-
-	if (!fileStats.isFile()) {
-		return `Invalid patch path (not a file): ${patchFilePath}`;
-	}
-
-	if (!isSupportedPatchFile(patchFilePath)) {
-		return `Unsupported patch extension. Use .ptm, .zip, or .upd files: ${patchFilePath}`;
-	}
-
-	try {
-		const server = ServersConfig.getCurrentServer();
-		const validationResult: ValidResponse = await sendPatchValidateRequest(server, patchFilePath);
-
-		if (validationResult.error !== 0) {
-			isError.value = true;
-			return [
-				`Patch validation failed for ${vscode.workspace.asRelativePath(patchFilePath, false)}.`,
-				`- Code: ${validationResult.errorCode}`,
-				`- Message: ${validationResult.message || "Unknown validation error."}`,
-				"",
-				"Check the selected server/environment and patch compatibility before retrying."
-			].join("\n");
-		}
-
-		return `Patch validated successfully: ${vscode.workspace.asRelativePath(patchFilePath, false)}.`;
-	} catch (error) {
-		isError.value = true;
-		return `Error during patch validation: ${error instanceof Error ? error.message : String(error)}`;
-	}
-}
-
-async function handleApplyPatchCommand(
-	prompt: string | undefined,
-	token: vscode.CancellationToken,
-	isError: { value: boolean },
-	applyOld: boolean
-): Promise<string> {
-	isError.value = true;
-
-	if (token.isCancellationRequested) {
-		return "Apply patch command canceled by user.";
-	}
-
-	const patchFromPrompt: string | undefined = prompt?.trim() ? extractPatchFileFromPrompt(prompt.trim()) : undefined;
-	if (!patchFromPrompt) {
-		return "Patch file is required. Usage: /apply-patch patches/fix.ptm or /apply-patch c:/patches/fix.zip";
-	}
-
-	const patchFilePath: string | undefined = resolvePatchFilePath(patchFromPrompt);
-	if (!patchFilePath) {
-		return `Unable to resolve patch file path: ${patchFromPrompt}`;
-	}
-
-	if (!fs.existsSync(patchFilePath)) {
-		return `Patch file not found: ${patchFilePath}`;
-	}
-
-	let fileStats: fs.Stats;
-	try {
-		fileStats = fs.statSync(patchFilePath);
-	} catch (error) {
-		return `Unable to access patch file: ${error instanceof Error ? error.message : String(error)}`;
-	}
-
-	if (!fileStats.isFile()) {
-		return `Invalid patch path (not a file): ${patchFilePath}`;
-	}
-
-	if (!isSupportedPatchFile(patchFilePath)) {
-		return `Unsupported patch extension. Use .ptm, .zip, or .upd files: ${patchFilePath}`;
-	}
-
-	try {
-		const server = ServersConfig.getCurrentServer();
-		const validationResult: ValidResponse = await sendPatchApplyRequest(server, patchFilePath, applyOld);
-
-		if (validationResult.error !== 0) {
-			isError.value = true;
-			return [
-				`Patch validation failed for ${vscode.workspace.asRelativePath(patchFilePath, false)}.`,
-				`- Code: ${validationResult.errorCode}`,
-				`- Message: ${validationResult.message || "Unknown validation error."}`,
-				"",
-				"Check the selected server/environment and patch compatibility before retrying."
-			].join("\n");
-		}
-
-		return `Patch validated successfully: ${vscode.workspace.asRelativePath(patchFilePath, false)}.`;
-	} catch (error) {
-		isError.value = true;
-		return `Error during patch validation: ${error instanceof Error ? error.message : String(error)}`;
 	}
 }
 
@@ -367,7 +158,7 @@ export async function waitForDiagnosticsChange(
 	token: vscode.CancellationToken,
 	timeoutMs: number = DIAGNOSTIC_WAIT_TIMEOUT_MS
 ): Promise<DiagnosticsWaitResult> {
-	console.log(`[tds-chat-tools] waitForDiagnosticsChange: start timeoutMs=${timeoutMs}`);
+	console.log(`[tds-chat-tools] [tds-chat-tools] waitForDiagnosticsChange: start timeoutMs=${timeoutMs}`);
 	return await new Promise<DiagnosticsWaitResult>((resolve) => {
 		let finished: boolean = false;
 		let idleTimer: NodeJS.Timeout | undefined;
@@ -380,7 +171,7 @@ export async function waitForDiagnosticsChange(
 			if (idleTimer) {
 				clearTimeout(idleTimer);
 			}
-			console.log("[tds-chat-tools] waitForDiagnosticsChange: diagnostics changed, scheduling idle resolve.");
+			console.log("[tds-chat-tools] [tds-chat-tools] waitForDiagnosticsChange: diagnostics changed, scheduling idle resolve.");
 
 			idleTimer = setTimeout(() => finish(), DIAGNOSTIC_IDLE_WINDOW_MS);
 		};
@@ -398,7 +189,7 @@ export async function waitForDiagnosticsChange(
 				clearTimeout(idleTimer);
 			}
 			clearTimeout(timeout);
-			console.log(`[tds-chat-tools] waitForDiagnosticsChange: finish changed=${hasChange} timedOut=${timedOut} canceled=${canceled}`);
+			console.log(`[tds-chat-tools] [tds-chat-tools] waitForDiagnosticsChange: finish changed=${hasChange} timedOut=${timedOut} canceled=${canceled}`);
 			resolve({
 				changed: hasChange,
 				timedOut,
@@ -476,7 +267,7 @@ function buildHelpText(): string {
 		vscode.l10n.t("- /help: shows this help."),
 		"",
 		vscode.l10n.t("Supported targets:"),
-		vscode.l10n.t("- current editor: editor, current"),
+		vscode.l10n.t("- current editor: editor, current (default)"),
 		vscode.l10n.t("- open editors: open-files, open-editors, arquivos-abertos, fontes-abertos, arquivos abertos, fontes abertos"),
 		vscode.l10n.t("- workspace: workspace, ws, all"),
 		vscode.l10n.t("- file: relative or absolute path (ex: src/meu.prw)"),
@@ -558,7 +349,7 @@ class ChatTdsTools implements vscode.LanguageModelTool<ChatCompilerToolInput> {
 		options: vscode.LanguageModelToolInvocationOptions<ChatCompilerToolInput>,
 		token: vscode.CancellationToken
 	): Promise<vscode.LanguageModelToolResult> {
-		console.log(`[tds-chat-tools] invoke: command='${options.input?.command ?? COMPILER_COMMAND}' target='${options.input?.target ?? "<default>"}' flags='${options.input?.flags?.join(",") ?? "none"}'`);
+		console.log(`[tds-chat-tools] [tds-chat-tools] invoke: command='${options.input?.command ?? COMPILER_COMMAND}' target='${options.input?.target ?? "<default>"}' flags='${options.input?.flags.applied}']`);
 
 		if (token.isCancellationRequested) {
 			console.warn("[tds-chat-tools] invoke: canceled before start.");
@@ -566,9 +357,8 @@ class ChatTdsTools implements vscode.LanguageModelTool<ChatCompilerToolInput> {
 				new vscode.LanguageModelTextPart("TDS tool invocation canceled by user.")
 			]);
 		}
-		const input: ChatCompilerToolInput = options.input ?? {};
-
-		const command: string | undefined = normalizeCommand(input.command);
+		const command: string | undefined = normalizeCommand(options.input?.command);
+		const _input: ChatCompilerToolInput = options.input ?? DEFAULT_COMPILER_TOOL_INPUT;
 		const isValid: string = isValidCommand(command, command);
 
 		if (isValid !== "") {
@@ -584,122 +374,38 @@ class ChatTdsTools implements vscode.LanguageModelTool<ChatCompilerToolInput> {
 			]);
 		}
 
-		const target: string = resolveTarget(input.target);
-		const targetUri: vscode.Uri | undefined = resolveTargetUri(input.target);
-		const flagOptions: FlagOptions = parseFlagOptions(input.flags);
-		const flags: string = flagOptions.applied.length > 0 ? flagOptions.applied.join(", ") : "none";
+		const target: string = resolveTarget(_input.target);
+		//const targetUri: vscode.Uri | undefined = resolveTargetUri(input.target);
+		const flagOptions: FlagOptions = _input.flags;  //parseFlagOptions(input.flags);
+		//const flags: string = flagOptions.applied.length > 0 ? flagOptions.applied.join(", ") : "none";
 
-		if ((input.command === VALIDATE_PATCH_COMMAND) || (input.command === APPLY_PATCH_COMMAND)) {
+		if ((_input.command === VALIDATE_PATCH_COMMAND)) {
 			const isError: { value: boolean } = { value: false };
-			const validationResult: string = await handleValidPatchCommand(input.target, token, isError);
+			const validationResult: string = await handleValidPatchCommand(_input.target, token, isError);
 			let finalMessage: string = !isError.value
-				? `Validating the patch for ${resolveTarget(input.target)}.`
+				? `Validating the patch for ${resolveTarget(_input.target)}.`
 				: validationResult;
 			return new vscode.LanguageModelToolResult([
 				new vscode.LanguageModelTextPart(finalMessage)
 			]);
 		}
 
-		if (input.command === APPLY_PATCH_COMMAND) {
-			const isError: { value: boolean } = { value: false };
-			const validationResult: string = await handleApplyPatchCommand(input.target, token, isError, flagOptions.applyOld);
-			if (isError.value) {
-				let finalMessage: string = !isError.value
-					? `Applying the patch for ${resolveTarget(input.target)}.`
-					: validationResult;
-				return new vscode.LanguageModelToolResult([
-					new vscode.LanguageModelTextPart(finalMessage)
-				]);
-			}
-		}
+		const isError: { value: boolean } = { value: false };
 
-		const isOpenEditorsTarget: boolean = target === "open-editors";
-		const isWorkspaceTarget: boolean = target === "workspace" && !targetUri;
-		const commandIdBase: string = isWorkspaceTarget
-			? REBUILD_WORKSPACE_COMMAND
-			: isOpenEditorsTarget
-				? REBUILD_OPEN_EDITORS_COMMAND
-				: REBUILD_FILE_COMMAND;
-		const commandId: string = flagOptions.syntaxOnly ? SYNTAX_ONLY_FILE_COMMAND : commandIdBase;
-		console.log(`[tds-chat-tools] invoke: resolved target='${target}', commandId='${commandId}', targetUri='${targetUri?.toString(true) ?? "<none>"}'`);
+		if (_input.command === APPLY_PATCH_COMMAND) {
+			const validationResult: string = await handleApplyPatchCommand(_input.target, token, isError, flagOptions.applyOld);
+			let finalMessage: string = !isError.value
+				? `Applying the patch for ${resolveTarget(_input.target)}.`
+				: validationResult;
 
-		try {
-			if (isWorkspaceTarget || isOpenEditorsTarget) {
-				console.log(`[tds-chat-tools] invoke: executing command '${commandId}' without explicit URI.`);
-				await vscode.commands.executeCommand(commandId);
-			} else if (targetUri) {
-				console.log(`[tds-chat-tools] invoke: executing command '${commandId}' with URI '${targetUri.toString(true)}'.`);
-				await vscode.commands.executeCommand(commandId, undefined, [targetUri]);
-			} else {
-				console.log(`[tds-chat-tools] invoke: executing command '${commandId}' fallback without URI.`);
-				await vscode.commands.executeCommand(commandId);
-			}
-		} catch (error) {
-			console.error(`[tds-chat-tools] invoke: command failed (${commandId}): ${error instanceof Error ? error.stack ?? error.message : String(error)}`);
 			return new vscode.LanguageModelToolResult([
-				new vscode.LanguageModelTextPart(
-					`TDS tool command failed (${commandId}): ${error instanceof Error ? error.message : String(error)}`
-				)
+				new vscode.LanguageModelTextPart(finalMessage)
 			]);
 		}
 
-		const diagnosticsTimeout: number = isDirectoryUri(targetUri)
-			? DIAGNOSTIC_WAIT_TIMEOUT_MS_FOLDER
-			: DIAGNOSTIC_WAIT_TIMEOUT_MS;
-		console.log(`[tds-chat-tools] invoke: waiting diagnostics with timeout=${diagnosticsTimeout}`);
-		const waitResult: DiagnosticsWaitResult = await waitForDiagnosticsChange(token, diagnosticsTimeout);
-		if (waitResult.canceled) {
-			console.warn("[tds-chat-tools] invoke: canceled while waiting diagnostics.");
-			return new vscode.LanguageModelToolResult([
-				new vscode.LanguageModelTextPart("Compiler tool invocation canceled by user.")
-			]);
-		}
-
-		const diagnostics: DiagnosticEntry[] = applyDiagnosticFilterOptions(
-			isOpenEditorsTarget ? collectDiagnosticsForOpenEditors() : collectDiagnostics(targetUri, isWorkspaceTarget),
-			flagOptions
-		);
-		const showAllDiagnostics: boolean = isDirectoryUri(targetUri);
-		const maxToShow: number = flagOptions.max && flagOptions.max > 0
-			? flagOptions.max
-			: (showAllDiagnostics ? diagnostics.length : MAX_DIAGNOSTICS_TO_SHOW);
-		const displayedDiagnostics: DiagnosticEntry[] = diagnostics.slice(0, maxToShow);
-		console.log(`[tds-chat-tools] invoke: diagnostics total=${diagnostics.length} displayed=${displayedDiagnostics.length} timedOut=${waitResult.timedOut} changed=${waitResult.changed}`);
-		const diagnosticsSummary: string = formatDiagnosticsSummary(
-			displayedDiagnostics,
-			target,
-			diagnostics.length > displayedDiagnostics.length,
-			flagOptions.format
-		);
-
-		const summary: string = flagOptions.format === "json"
-			? JSON.stringify({
-				command: commandId,
-				target,
-				flags: flags,
-				diagnosticsUpdated: waitResult.changed,
-				timedOut: waitResult.timedOut,
-				diagnostics: JSON.parse(diagnosticsSummary)
-			}, null, 2)
-			: [
-				"Compiler tool executed successfully with:",
-				//`- command: ${commandId}`,
-				`- target: ${target}`,
-				`- compiler`,
-				`- flags: ${flags}`,
-				//`- diagnostics-updated: ${waitResult.changed ? "true" : "false"}`,
-				"",
-				...(waitResult.timedOut && !waitResult.changed
-					? [
-						"Diagnostics were not updated before timeout. Results below may reflect a previous compile.",
-						""
-					]
-					: []),
-				diagnosticsSummary
-			].join("\n");
-
+		const validationResult: string[] = await handleCompilerCommand(target, flagOptions, token, isError);
 		return new vscode.LanguageModelToolResult([
-			new vscode.LanguageModelTextPart(summary)
+			new vscode.LanguageModelTextPart(validationResult.join("\n"))
 		]);
 	}
 }
@@ -718,11 +424,11 @@ async function compilerParticipantHandler(
 	stream: vscode.ChatResponseStream,
 	token: vscode.CancellationToken
 ): Promise<vscode.ChatResult | void> {
-
 	const rawPrompt: string | undefined = request.prompt?.trim();
 	let commandPrompt: string | undefined = rawPrompt;
 	let command: string | undefined = normalizeCommand(request.command ?? commandPrompt);
-	console.log(`[tds-chat-tools] participant: received request prompt='${rawPrompt ?? ""}' command='${command ?? ""}'`);
+
+	console.log(`[tds-chat-tools] [tds-chat-tools] participant: received request prompt='${rawPrompt ?? ""}' command='${command ?? ""}'`);
 	const isValid: string = isValidCommand(command, commandPrompt);
 
 	if (isValid !== "") {
@@ -751,64 +457,52 @@ async function compilerParticipantHandler(
 	const isSyntaxOnly: boolean = command === SYNTAX_ONLY_COMMAND;
 	const isApplyPatch: boolean = command === APPLY_PATCH_COMMAND;
 	const isValidatePatch: boolean = command === VALIDATE_PATCH_COMMAND;
-	const input: ChatCompilerToolInput = {
-		command,
-		target: resolveTarget(),
-		//syntaxOnly: isSyntaxOnly
-	};
+	const compilerToolInput: ChatCompilerToolInput = parserPrompt(command, commandPrompt);
 
+	// if (isApplyPatch) {
+	// 	input.target = commandPrompt;
+	// 	delete input.target;
+	// 	//delete input.syntaxOnly;
+	// }
+
+	// if (!isApplyPatch && commandPrompt?.trim()) {
+	// 	if (isSyntaxOnly && !compilerToolInput) {
+	// 		flags.push("only=all");
+	// 	}
+
+	// 	input.flags = flags;
+	// } else if (!isApplyPatch && isSyntaxOnly) {
+	// 	input.flags = ["only=all"];
+	// }
+	console.log(`[tds-chat-tools] [tds-chat-tools] participant: invoking tool command='${compilerToolInput.command ?? ""}' target='${compilerToolInput.target ?? "<default>"}' flags='${compilerToolInput.flags.applied}'`);
+
+	const resolvedTarget: string = resolveTarget(compilerToolInput.target);
 	if (isApplyPatch) {
-		input.target = commandPrompt;
-		delete input.target;
-		//delete input.syntaxOnly;
-	}
-
-	if (!isApplyPatch && commandPrompt?.trim()) {
-		const prompt: string = commandPrompt.trim();
-		const targetFromPrompt: string | undefined = extractTargetFromPrompt(prompt);
-		if (targetFromPrompt) {
-			input.target = targetFromPrompt;
-		}
-
-		const flags: string[] = [prompt];
-		const hasOnlyFlag: boolean = /(?:^|\s)only\s*[:=]/i.test(prompt);
-		if (isSyntaxOnly && !hasOnlyFlag) {
-			flags.push("only=all");
-		}
-
-		input.flags = flags;
-	} else if (!isApplyPatch && isSyntaxOnly) {
-		input.flags = ["only=all"];
-	}
-	console.log(`[tds-chat-tools] participant: invoking tool command='${input.command ?? ""}' target='${input.target ?? "<default>"}' flags='${input.flags?.join(",") ?? "none"}'`);
-
-	const resolvedTarget: string = resolveTarget(input.target);
-	if (isApplyPatch) {
-		stream.markdown("Validating patch package...\n\n");
+		console.log("[tds-chat-tools] Validating patch package...\n\n");
 	} else if (isSyntaxOnly) {
-		stream.markdown(`Starting syntax-only compilation for ${resolvedTarget}...\n\n`);
+		console.log(`[tds-chat-tools] Starting syntax-only compilation for ${resolvedTarget}...\n\n`);
 	} else if (isValidatePatch) {
-		stream.markdown("Validating patch...\n\n");
+		console.log("[tds-chat-tools] Validating patch...\n\n");
 	} else {
-		stream.markdown(`Starting compilation for ${resolvedTarget}...\n\n`);
+		console.log(`[tds-chat-tools] Starting compilation for ${resolvedTarget}...\n\n`);
 	}
 
 	const result: vscode.LanguageModelToolResult = await vscode.lm.invokeTool(
 		COMPILER_TOOL_NAME,
 		{
-			input,
+			input: compilerToolInput,
 			toolInvocationToken: request.toolInvocationToken
 		},
 		token
 	);
-	console.log(`[tds-chat-tools] participant: tool invocation finished for target='${resolvedTarget}'`);
+	console.log(`[tds-chat-tools] [tds-chat-tools] participant: tool invocation finished for target='${resolvedTarget}'`);
 
 	if (isApplyPatch) {
-		stream.markdown(`Patching complete for ${resolvedTarget}.\n\n`);
+		console.log(`[tds-chat-tools] Patching complete for ${resolvedTarget}.\n\n`);
 	} else if (isValidatePatch) {
-		stream.markdown(`Patch validation complete for ${resolvedTarget}.\n\n`);
+		console.log(`[tds-chat-tools] Patch validation complete for ${resolvedTarget}.\n\n`);
 	} else {
-		stream.markdown(`Compilation finished for ${resolvedTarget}.\n\n`);
+		console.log(`[tds-chat-tools] Compilation finished for ${resolvedTarget}.\n\n`);
 	}
 
 	const text: string = result.content
@@ -817,7 +511,7 @@ async function compilerParticipantHandler(
 		.join("\n");
 
 	stream.markdown(text || "Compiler tool invoked successfully.");
-	console.log(`[tds-chat-tools] participant: streamed result length=${text.length}`);
+	console.log(`[tds-chat-tools] [tds-chat-tools] participant: streamed result length=${text.length}`);
 
 	return {
 		metadata: {
@@ -848,4 +542,42 @@ export function registerChatTools(context: vscode.ExtensionContext): vscode.Disp
 
 	disposables.push(participant);
 	return disposables;
+}
+
+function parserPrompt(command: string, prompt: string): ChatCompilerToolInput {
+	const result: ChatCompilerToolInput = {
+		command: command,
+		target: "",
+		flags: structuredClone(DEFAULT_FLAG_OPTIONS)
+	};
+	const parts: string[] = prompt.split(/\s+/);
+	let isFirstPart: boolean = true;
+
+		for (const part of parts) {
+			const subPart: string[] = part.split(/[=\:]/);
+
+			if (subPart.length === 2) {
+				const key: string = subPart[0].toLowerCase();
+				const value: string = subPart[1];
+
+				if (key === "target") {
+					result.target = value;
+				} else {
+					parseFlagOptions(result.flags, key, value);
+				}
+			} else if (subPart.length === 1 && isFirstPart) {
+				result.target = subPart[0];
+			}
+			isFirstPart = false;
+		}
+
+	if (result.target.length === 0) {
+		result.target = "current";
+	}
+
+	if (result.flags.applied.length === 0) {
+		result.flags.applied.push("(none)");
+	}
+
+	return result;
 }
