+++ src/chat/chatUtils.ts
@@ -35,9 +35,24 @@ export const OPEN_EDITORS_TARGET_ALIASES: Set<string> = new Set([
 ]);
 
 export type ChatCompilerToolInput = {
-	command?: string;
-	target?: string;
-	flags?: string[];
+	command: string;
+	target: string;
+	flags: FlagOptions;
+};
+
+export const DEFAULT_FLAG_OPTIONS: FlagOptions = {
+	only: "all",
+	sort: "none",
+	format: "markdown",
+	syntaxOnly: false,
+	applyOld: false,
+	applied: []
+};
+
+export const DEFAULT_COMPILER_TOOL_INPUT: ChatCompilerToolInput = {
+	command: COMPILER_COMMAND,
+	target: "",
+	flags: { ...DEFAULT_FLAG_OPTIONS }
 };
 
 export type DiagnosticEntry = {
@@ -102,7 +117,7 @@ export function normalizeCommand(value: string): string {
  * @returns Target normalized to lowercase.
  */
 export function normalizeTargetKeyword(value: string): string {
-	return value.trim().toLowerCase();
+	return stripQuotes(value).trim().toLowerCase();
 }
 
 /**
@@ -238,123 +253,6 @@ export function isDirectoryUri(uri: vscode.Uri | undefined): boolean {
 }
 
 /**
- * Collects error/warning diagnostics for a file, folder, or workspace.
- * @param targetUri Target URI for build.
- * @param isWorkspaceTarget Indicates whether the build was requested for the entire workspace.
- * @returns List of diagnostics eligible for chat display.
- */
-export function collectDiagnostics(targetUri: vscode.Uri | undefined, isWorkspaceTarget: boolean): DiagnosticEntry[] {
-	const diagnosticsByUri: [vscode.Uri, vscode.Diagnostic[]][] = vscode.languages.getDiagnostics();
-	let filtered: [vscode.Uri, vscode.Diagnostic[]][] = diagnosticsByUri;
-
-	if (!isWorkspaceTarget && targetUri) {
-		if (isDirectoryUri(targetUri)) {
-			filtered = diagnosticsByUri.filter(([uri]) => isUriInsideFolder(targetUri, uri));
-		} else {
-			filtered = diagnosticsByUri.filter(([uri]) => uriEquals(uri, targetUri));
-		}
-
-		if (filtered.length === 0) {
-			filtered = diagnosticsByUri.filter(([uri]) => isWorkspaceUri(uri));
-		}
-	} else {
-		filtered = diagnosticsByUri.filter(([uri]) => isWorkspaceUri(uri));
-	}
-
-	const entries: DiagnosticEntry[] = [];
-	for (const [uri, diagnostics] of filtered) {
-		for (const diagnostic of diagnostics) {
-			if (
-				diagnostic.severity === vscode.DiagnosticSeverity.Error ||
-				diagnostic.severity === vscode.DiagnosticSeverity.Warning
-			) {
-				entries.push({ uri, diagnostic });
-			}
-		}
-	}
-
-	return entries;
-}
-
-
-/**
- * Applies diagnostic filters and sorting based on provided flags.
- * @param entries Collected diagnostics.
- * @param options Filter/sort options.
- * @returns Filtered and sorted diagnostics.
- */
-export function applyDiagnosticFilterOptions(
-	entries: DiagnosticEntry[],
-	options: FlagOptions
-): DiagnosticEntry[] {
-	let result: DiagnosticEntry[] = entries.slice();
-
-	if (options.only === "error") {
-		result = result.filter((entry) => entry.diagnostic.severity === vscode.DiagnosticSeverity.Error);
-	} else if (options.only === "warning") {
-		result = result.filter((entry) => entry.diagnostic.severity === vscode.DiagnosticSeverity.Warning);
-	}
-
-	if (options.sort === "file") {
-		result.sort((a, b) => {
-			const pathA: string = a.uri.scheme === "file" ? a.uri.fsPath.toLowerCase() : a.uri.toString(true).toLowerCase();
-			const pathB: string = b.uri.scheme === "file" ? b.uri.fsPath.toLowerCase() : b.uri.toString(true).toLowerCase();
-			if (pathA < pathB) {
-				return -1;
-			}
-			if (pathA > pathB) {
-				return 1;
-			}
-
-			const lineA: number = a.diagnostic.range.start.line;
-			const lineB: number = b.diagnostic.range.start.line;
-			return lineA - lineB;
-		});
-	} else if (options.sort === "severity") {
-		result.sort((a, b) => {
-			const severityA: vscode.DiagnosticSeverity = a.diagnostic.severity ?? vscode.DiagnosticSeverity.Hint;
-			const severityB: vscode.DiagnosticSeverity = b.diagnostic.severity ?? vscode.DiagnosticSeverity.Hint;
-			if (severityA !== severityB) {
-				return severityA - severityB;
-			}
-
-			const pathA: string = a.uri.scheme === "file" ? a.uri.fsPath.toLowerCase() : a.uri.toString(true).toLowerCase();
-			const pathB: string = b.uri.scheme === "file" ? b.uri.fsPath.toLowerCase() : b.uri.toString(true).toLowerCase();
-			if (pathA < pathB) {
-				return -1;
-			}
-			if (pathA > pathB) {
-				return 1;
-			}
-
-			return a.diagnostic.range.start.line - b.diagnostic.range.start.line;
-		});
-	}
-
-	return result;
-}
-
-/**
- * Returns a standardized label for diagnostic severity.
- * @param severity Diagnostic severity.
- * @returns Standardized text for chat display.
- */
-export function severityLabel(severity: vscode.DiagnosticSeverity): string {
-	switch (severity) {
-		case vscode.DiagnosticSeverity.Error:
-			return "ERROR";
-		case vscode.DiagnosticSeverity.Warning:
-			return "WARN";
-		case vscode.DiagnosticSeverity.Information:
-			return "INFO";
-		case vscode.DiagnosticSeverity.Hint:
-			return "HINT";
-		default:
-			return "UNKNOWN";
-	}
-}
-
-/**
  * Checks whether a URI is inside a parent folder (recursively).
  * @param parentFolder Base folder URI.
  * @param targetUri Resource URI to validate.
@@ -398,102 +296,13 @@ export function uriEquals(left: vscode.Uri, right: vscode.Uri): boolean {
  * @param uri URI to validate.
  * @returns True when the URI belongs to the workspace.
  */
-function isWorkspaceUri(uri: vscode.Uri): boolean {
+export function isWorkspaceUri(uri: vscode.Uri): boolean {
 	if (uri.scheme !== "file") {
 		return false;
 	}
 
 	const folder: vscode.WorkspaceFolder | undefined = vscode.workspace.getWorkspaceFolder(uri);
 	return !!folder;
-}
-
-/**
- * Builds the final text with summary and diagnostics list for chat.
- * @param entries Diagnostics to display.
- * @param targetLabel Name/identifier of the compiled target.
- * @param truncated Indicates whether the list was truncated by limit.
- * @returns Final response text for chat.
- */
-export function formatDiagnosticsSummary(
-	entries: DiagnosticEntry[],
-	targetLabel: string,
-	truncated: boolean,
-	outputFormat: "markdown" | "json"
-): string {
-	const errors: number = entries.filter((entry) => entry.diagnostic.severity === vscode.DiagnosticSeverity.Error).length;
-	const warnings: number = entries.length - errors;
-
-	if (outputFormat === "json") {
-		const diagnostics = entries.map((entry) => {
-			const relativePath: string =
-				entry.uri.scheme === "file"
-					? vscode.workspace.asRelativePath(entry.uri, false)
-					: entry.uri.toString(true);
-			const code: string | number | undefined =
-				typeof entry.diagnostic.code === "string"
-					? entry.diagnostic.code
-					: entry.diagnostic.code && typeof entry.diagnostic.code === "object"
-						? entry.diagnostic.code.value
-						: undefined;
-			return {
-				severity: severityLabel(entry.diagnostic.severity),
-				message: entry.diagnostic.message.replace(/\s+/g, " ").trim(),
-				source: entry.diagnostic.source ?? null,
-				code: code ?? null,
-				line: entry.diagnostic.range.start.line + 1,
-				path: relativePath,
-				uri: entry.uri.toString(true)
-			};
-		});
-
-		return JSON.stringify({
-			target: targetLabel,
-			errors,
-			warnings,
-			truncated,
-			diagnostics
-		}, null, 2);
-	}
-
-	if (entries.length === 0) {
-		return [
-			"Compilation completed and no errors or warnings were found in Problems.",
-		].join("\n");
-	}
-	const details: string[] = entries.map((entry, index) => {
-		const relativePath: string =
-			entry.uri.scheme === "file"
-				? vscode.workspace.asRelativePath(entry.uri, false)
-				: entry.uri.toString(true);
-		const line: number = entry.diagnostic.range.start.line + 1;
-		const locationLink: string = toLocationLink(entry.uri, relativePath, line);
-		const code: string | number | undefined =
-			typeof entry.diagnostic.code === "string"
-				? entry.diagnostic.code
-				: entry.diagnostic.code && typeof entry.diagnostic.code === "object"
-					? entry.diagnostic.code.value
-					: undefined;
-		const cleanMessage: string = entry.diagnostic.message.replace(/\s+/g, " ").trim();
-		const source: string = entry.diagnostic.source ? ` (${entry.diagnostic.source})` : "";
-		const codeText: string = code ? ` [${code}]` : "";
-
-		return `${index + 1}. ${severityLabel(entry.diagnostic.severity)} ${locationLink}${source}${codeText} - ${cleanMessage}`;
-	});
-
-	const lines: string[] = [
-		"Compilation finished with diagnostics:",
-		`- target: ${targetLabel}`,
-		`- errors: ${errors}`,
-		`- warnings: ${warnings}`,
-		"",
-		...details
-	];
-
-	if (truncated) {
-		lines.push("", `Showing first ${MAX_DIAGNOSTICS_TO_SHOW} diagnostics.`);
-	}
-
-	return lines.join("\n");
 }
 
 /**
@@ -510,4 +319,24 @@ export function toLocationLink(uri: vscode.Uri, relativePath: string, line: numb
 
 	const locationUri: string = uri.with({ fragment: `L${line}` }).toString(true);
 	return `[${relativePath}:${line}](${locationUri})`;
+}
+
+/**
+ * Returns a standardized label for diagnostic severity.
+ * @param severity Diagnostic severity.
+ * @returns Standardized text for chat display.
+ */
+export function severityLabel(severity: vscode.DiagnosticSeverity): string {
+	switch (severity) {
+		case vscode.DiagnosticSeverity.Error:
+			return "ERROR";
+		case vscode.DiagnosticSeverity.Warning:
+			return "WARN";
+		case vscode.DiagnosticSeverity.Information:
+			return "INFO";
+		case vscode.DiagnosticSeverity.Hint:
+			return "HINT";
+		default:
+			return "UNKNOWN";
+	}
 }
